+++ tools/stellar-archivist/archivist/dirprefix_test.go
